Frontend | Add files for components
