+++ frontend/src/components/Shared/Centered/Centered.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const Centered = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+`;
