+++ frontend/src/components/Shared/Centered/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Centered";
