+++ frontend/src/components/TournamentRegistrations/index.tsx
@@ -1,0 +1,1 @@
+export * from "./TournamentRegistrations";
